+++ Editor/Drawers/LayerDrawer/LayerStringDropdownElement.cs
@@ -1,5 +1,6 @@
 #if UNITY_2021_3_OR_NEWER
 using SaintsField.Editor.UIToolkitElements;
+using UnityEngine;
 using UnityEngine.UIElements;
 
 
@@ -16,7 +17,11 @@ namespace SaintsField.Editor.Drawers.LayerDrawer
                     return;
                 }
 
-                LayerUtils.MakeDropdown(false, _maskValue, root, newValue => value = newValue.Name);
+                LayerUtils.MakeDropdown(false, _maskValue, root, newValue =>
+                {
+                    // Debug.Log($"dropdown set value to {newValue.Name}");
+                    value = newValue.Name;
+                });
             };
         }
 
@@ -24,20 +29,30 @@ namespace SaintsField.Editor.Drawers.LayerDrawer
 
         public override void SetValueWithoutNotify(string newValue)
         {
+            // Debug.Log($"set value {newValue}");
             CachedValue = newValue;
-
             foreach (LayerUtils.LayerInfo layerInfo in LayerUtils.GetAllLayers())
             {
                 if (layerInfo.Name == newValue)
                 {
-                    Label.text = LayerUtils.LayerInfoLabelUIToolkit(layerInfo);
+                    // Debug.Log($"set label to {LayerUtils.LayerInfoLabelUIToolkit(layerInfo)}");
+                    // Label.text = LayerUtils.LayerInfoLabelUIToolkit(layerInfo);
+                    // ((INotifyValueChanged<string>)Label).SetValueWithoutNotify(LayerUtils.LayerInfoLabelUIToolkit(layerInfo));
+                    SetLabelString(LayerUtils.LayerInfoLabelUIToolkit(layerInfo));
                     _maskValue = layerInfo.Mask;
+                    Button.tooltip = layerInfo.Name;
                     return;
                 }
             }
 
-            Label.text = $"<color=red>?</color> {newValue}";
+            // Debug.Log($"set invalid label to question {newValue}");
+
+            // Label.text = $"<color=red>?</color> {newValue}";
+            // ((INotifyValueChanged<string>)Label).SetValueWithoutNotify($"<color=red>?</color> {newValue}");
+            SetLabelString($"<color=red>?</color> {newValue}");
+
             _maskValue = -1;
+            Button.tooltip = "Invalid layer";
         }
 
     }