+++ Samples~/Scripts/LayerAttributeExample.cs
@@ -5,7 +5,8 @@ namespace SaintsField.Samples.Scripts
 {
     public class LayerAttributeExample: SaintsMonoBehaviour
     {
-        [Layer][FieldLabelText("<icon=star.png /><label />")] public string layerString;
+        [LabelText("<icon=star.png /><label />")]
+        [Layer] public string layerString;
         [Layer] public int layerInt;
 
         [ReadOnly]
@@ -15,5 +16,29 @@ namespace SaintsField.Samples.Scripts
 
         [Layer]
         public LayerMask lm;
+
+        [ShowInInspector] private LayerMask myLayerMaskRaw
+        {
+            get => myLayerMask;
+            set => myLayerMask = value;
+        }
+
+        [ShowInInspector, Layer] private LayerMask lmRaw
+        {
+            get => lm;
+            set => lm = value;
+        }
+
+        [ShowInInspector, Layer] private int layerIntRaw
+        {
+            get => layerInt;
+            set => layerInt = value;
+        }
+
+        [ShowInInspector, Layer] private string layerStringRaw
+        {
+            get => layerString;
+            set => layerString = value;
+        }
     }
 }